Use named pathfinding constants.
- Rename CostForInvalidCell to PathCostForInvalidPath
- Add MovementCostForUnreachableCell
- Update usages of int.MaxValue and short.Maxvalue to use named constants where relevant.
- Update costs on ICustomMovementLayer to return short, for consistency with costs from Locomotor.
- Rename some methods to distinguish between path/movement cost.

+++ OpenRA.Mods.Common/Traits/World/JumpjetLocomotor.cs
@@ -17,7 +17,7 @@ namespace OpenRA.Mods.Common.Traits
 	public class JumpjetLocomotorInfo : LocomotorInfo
 	{
 		[Desc("Pathfinding cost for taking off or landing.")]
-		public readonly int JumpjetTransitionCost = 0;
+		public readonly short JumpjetTransitionCost = 0;
 
 		[Desc("The terrain types that this actor can transition on. Leave empty to allow any.")]
 		public readonly HashSet<string> JumpjetTransitionTerrainTypes = new HashSet<string>();
